src/app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
requests
pillow
python-dotenv
passlib[bcrypt]
pydantic